[1.3.1] Merged revision(s) 4621 from trunk: Updated file writing

+++ branches/1.3.1/MTA10/mods/shared_logic/CClientPed.cpp
@@ -3482,7 +3482,7 @@ void CClientPed::_ChangeModel ( void )
                 m_pPlayerPed->RemoveGeometryRef ();
             #endif
                 // As we will have problem removing the geometry later, we might as well keep the model cached until exit
-                g_pCore->AddModelToPersistentCache ( m_ulModel );
+                g_pCore->AddModelToPersistentCache ( (ushort)m_ulModel );
 
             }
